+++ WhatsNew.md
@@ -2,6 +2,7 @@
 
 ## Functionality
 
++ Show date of next opposition for outer planets and for next greatest eastern / western elongation for innet planets.
 + A rich text editor can be used for some fields, like the description of an equipment set, the observations, the description of an observing list, ...
 + New menu structure
 + Improved responsivity of DeepskyLog by not reloading the whole page every time.